Update scod2.tlk
Made all the classes (base, prestige, racials, and templates) have the standard Gold color (Most already had this, some did not) as well as adding the Not-Character-Creator red line for prestige classes to all prestige, racial, and template classes.

Also, for Palemaster and Palemaster(Warlock) I changed the TLK to show the requirements for both, as the 2da entries both point to the same TLK StrRef.  If you wind up not liking the look, the better way would be to edit the 2da for one of those entries to point to a brand new StrRef so they can be properly seperate StrRefs (Such as was previously done properly with Arcane Trickster and its warlock subsidiary).