Custom Materials to ID harmful pots
A potion of harm or inflict cannot normally be used in no pvp areas because of the SetHostile flag, so a workaround is to make them use UniquePowerSelfOnly instead of CastSpell, but then if a player changes the name and desc the potion becomes a mystery to other players. So I added special materials that simply correspond to the names of those 6 spells and add them on those potions as the base material.