Dec Update
See oing it in exactly that order results in no compiler errors or even any compiler warnings.

Should be stable enough to go to main but I can't promise there's zero bugs.  As i've mentioned before, I typically only have 1 day a week to work on this stuff which doesn't leave a massive amount of time for testing. Anyway, that's all. You can just share the link to this google doc for the announcement

https://docs.google.com/document/d/1gZUAm5Ufb8CaEAWwdSLkJizckmObDNy30i4enys_Jl8/edit?usp=sharing