Change to Energy Immunity
Energy Immunity was Innate Level 7, Wiz level 7, and Cleric level 6. This meant that a cleric/FS or druid/shaman with 30 ranks of craft alch could make a potion of energy immune but a wiz could not. Lowered Innate lvl to 6 so that a wiz CAN. Wiz will still get the spell at lvl 7, though.